+++ src/index.js
@@ -1,10 +1,10 @@
-import * as content from "./layouts/sample/content";
+import * as content from "./layouts/sample";
 import './index.css';
 
 content.render();
 
 if (module.hot) {
-  module.hot.accept('./layouts/sample/content', () => {
+  module.hot.accept('./layouts/sample/index', () => {
     content.clean();
     content.render();
   });
